Fix Interaction Core V2 merge blockers

+++ Ayane/Models/PortableJSONValue.swift
@@ -7,7 +7,7 @@ import Foundation
 /// fields created by another compatible frontend during import/export.
 enum PortableJSONValue: Codable, Equatable, Sendable {
     case string(String)
-    case number(Double)
+    case number(Decimal)
     case bool(Bool)
     case object([String: PortableJSONValue])
     case array([PortableJSONValue])
@@ -19,7 +19,7 @@ enum PortableJSONValue: Codable, Equatable, Sendable {
             self = .null
         } else if let value = try? container.decode(Bool.self) {
             self = .bool(value)
-        } else if let value = try? container.decode(Double.self) {
+        } else if let value = try? container.decode(Decimal.self) {
             self = .number(value)
         } else if let value = try? container.decode(String.self) {
             self = .string(value)
